requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
psycopg2-binary==2.9.3
djangorestframework-simplejwt